Move subviewmodel creation to assembly
