+++ Teacher/Teacher/SpeedGrader/SpeedGraderAssembly.swift
@@ -16,8 +16,10 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //
 
+import Combine
 import Core
 import UIKit
+import SwiftUI
 
 enum SpeedGraderAssembly {
 
@@ -54,6 +56,67 @@ enum SpeedGraderAssembly {
             viewModel: viewModel
         )
         return CoreHostingController(view)
+    }
+
+    static func makePageViewModel(
+        assignment: Assignment,
+        submission: Submission,
+        contextColor: AnyPublisher<Color, Never>,
+        gradeStatusInteractor: GradeStatusInteractor,
+        submissionWordCountInteractor: SubmissionWordCountInteractor,
+        customGradebookColumnsInteractor: CustomGradebookColumnsInteractor,
+        env: AppEnvironment
+    ) -> SpeedGraderPageViewModel {
+        let rubricGradingInteractor = RubricGradingInteractorLive(
+            assignment: assignment,
+            submission: submission
+        )
+
+        let gradeInteractor = GradeInteractorLive(
+            assignment: assignment,
+            submission: submission,
+            rubricGradingInteractor: rubricGradingInteractor,
+            env: env
+        )
+
+        return SpeedGraderPageViewModel(
+            assignment: assignment,
+            latestSubmission: submission,
+            contextColor: contextColor,
+            studentAnnotationViewModel: .init(submission: submission),
+            gradeViewModel: .init(
+                assignment: assignment,
+                submission: submission,
+                gradeInteractor: gradeInteractor
+            ),
+            gradeStatusViewModel: .init(
+                userId: submission.userID,
+                submissionId: submission.id,
+                attempt: submission.attempt,
+                interactor: gradeStatusInteractor
+            ),
+            commentListViewModel: SubmissionCommentsAssembly.makeCommentListViewModel(
+                assignment: assignment,
+                latestSubmission: submission,
+                latestAttemptNumber: submission.attempt,
+                contextColor: contextColor,
+                env: env
+            ),
+            rubricsViewModel: .init(
+                assignment: assignment,
+                submission: submission,
+                interactor: rubricGradingInteractor
+            ),
+            submissionWordCountViewModel: .init(
+                userId: submission.userID,
+                attempt: submission.attempt,
+                interactor: submissionWordCountInteractor
+            ),
+            studentNotesViewModel: .init(
+                userId: submission.userID,
+                interactor: customGradebookColumnsInteractor
+            )
+        )
     }
 
 #if DEBUG
