+++ Teacher/Teacher/SpeedGrader/SpeedGraderScreen/ViewModel/SpeedGraderScreenViewModel.swift
@@ -180,39 +180,16 @@ extension SpeedGraderScreenViewModel: PagesViewControllerDataSource {
 
         let assignment = data.assignment
         let submission = data.submissions[index]
-        let rubricGradingInteractor = RubricGradingInteractorLive(assignment: assignment, submission: submission)
-
-        let rubricsViewModel = RubricsViewModel(
-            assignment: assignment,
-            submission: submission,
-            interactor: rubricGradingInteractor
-        )
-
-        let gradeInteractor = GradeInteractorLive(
-            assignment: assignment,
-            submission: submission,
-            rubricGradingInteractor: rubricGradingInteractor,
-            env: environment
-        )
-
-        let gradeViewModel = SpeedGraderSubmissionGradesViewModel(
-            assignment: assignment,
-            submission: submission,
-            gradeInteractor: gradeInteractor
-        )
 
         return SpeedGraderPageView(
-            env: environment,
             userIndexInSubmissionList: index,
-            viewModel: SpeedGraderPageViewModel(
+            viewModel: SpeedGraderAssembly.makePageViewModel(
                 assignment: assignment,
-                latestSubmission: submission,
+                submission: submission,
                 contextColor: interactor.contextInfo.compactMap { $0?.courseColor }.eraseToAnyPublisher(),
                 gradeStatusInteractor: interactor.gradeStatusInteractor,
                 submissionWordCountInteractor: interactor.submissionWordCountInteractor,
                 customGradebookColumnsInteractor: interactor.customGradebookColumnsInteractor,
-                rubricsViewModel: rubricsViewModel,
-                gradeViewModel: gradeViewModel,
                 env: environment
             ),
             landscapeSplitLayoutViewModel: landscapeSplitLayoutViewModel,
