+++ Teacher/TeacherTests/SpeedGrader/Comments/ViewModel/SubmissionCommentListViewModelTests.swift
@@ -348,69 +348,3 @@ class SubmissionCommentListViewModelTests: TeacherTestCase {
         )
     }
 }
-
-private final class SubmissionCommentsInteractorMock: SubmissionCommentsInteractor {
-
-    // MARK: - getSubmissionAttempts
-
-    var getSubmissionAttemptsCallsCount: Int = 0
-    var getSubmissionAttemptsResult: AnyPublisher<[Submission], Error>?
-    func getSubmissionAttempts() -> AnyPublisher<[Submission], Error> {
-        getSubmissionAttemptsCallsCount += 1
-        return getSubmissionAttemptsResult ?? Publishers.typedJust([])
-    }
-
-    // MARK: - getComments
-
-    var getCommentsCallsCount: Int = 0
-    var getCommentsResult: AnyPublisher<[SubmissionComment], Error>?
-    func getComments() -> AnyPublisher<[SubmissionComment], Error> {
-        getCommentsCallsCount += 1
-        return getCommentsResult ?? Publishers.typedJust([])
-    }
-
-    // MARK: - getIsAssignmentEnhancementsEnabled
-
-    var getIsAssignmentEnhancementsEnabledCallsCount: Int = 0
-    var getIsAssignmentEnhancementsEnabledResult: AnyPublisher<Bool, Error>?
-    func getIsAssignmentEnhancementsEnabled() -> AnyPublisher<Bool, Error> {
-        getIsAssignmentEnhancementsEnabledCallsCount += 1
-        return getIsAssignmentEnhancementsEnabledResult ?? Publishers.typedJust(false)
-    }
-
-    // MARK: - getIsCommentLibraryEnabled
-
-    var getIsCommentLibraryEnabledCallsCount: Int = 0
-    var getIsCommentLibraryEnabledResult: AnyPublisher<Bool, Error>?
-    func getIsCommentLibraryEnabled() -> AnyPublisher<Bool, Error> {
-        getIsCommentLibraryEnabledCallsCount += 1
-        return getIsCommentLibraryEnabledResult ?? Publishers.typedJust(false)
-    }
-
-    // MARK: - createTextComment
-
-    var createTextCommentCallsCount: Int = 0
-    var createTextCommentInput: (text: String, attemptNumber: Int?, completion: (Result<Void, Error>) -> Void)?
-    func createTextComment(_ text: String, attemptNumber: Int?, completion: @escaping (Result<Void, Error>) -> Void) {
-        createTextCommentCallsCount += 1
-        createTextCommentInput = (text: text, attemptNumber: attemptNumber, completion: completion)
-    }
-
-    // MARK: - createMediaComment
-
-    var createMediaCommentCallsCount: Int = 0
-    var createMediaCommentInput: (type: MediaCommentType, url: URL, attemptNumber: Int?, completion: (Result<Void, Error>) -> Void)?
-    func createMediaComment(type: MediaCommentType, url: URL, attemptNumber: Int?, completion: @escaping (Result<Void, Error>) -> Void) {
-        createMediaCommentCallsCount += 1
-        createMediaCommentInput = (type: type, url: url, attemptNumber: attemptNumber, completion: completion)
-    }
-
-    // MARK: - createFileComment
-
-    var createFileCommentCallsCount: Int = 0
-    var createFileCommentInput: (batchId: String, attemptNumber: Int?, completion: (Result<Void, Error>) -> Void)?
-    func createFileComment(batchId: String, attemptNumber: Int?, completion: @escaping (Result<Void, Error>) -> Void) {
-        createFileCommentCallsCount += 1
-        createFileCommentInput = (batchId: batchId, attemptNumber: attemptNumber, completion: completion)
-    }
-}
